cannot compare Strings with ==
general rule: you cannot compare Strings (for instance, "dragon") with ==, you ALWAYS need to use .equals.

because in Processing and Java, String are objects and with == you would compare the objects and not their contents. There can be multiple objects of type String that contain "dragon", that's why you need to use .equals, which compares the objects' content.